requirements.txt
fastapi
uvicorn

#for template
jinja2
python-jose
aiofiles
alembic

sqlalchemy
starlette
psycopg2-binary
pydantic
python-dotenv